+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - JSON writer and reader
 - Easier HTTP field setting option
 - Saving cookies to file
+- Quick way to construct a string from form parameters
 
 ### Fixed
 - Cookie share support